factored CPP source's main header into collection of per-source-file headers
